attacks/entra/apps/requirements.txt
faker
msgraph-sdk
azure-identity